Changes in function prototypes
I think "change" is the right word to encapsulate the additions and deletions I did on this code department. Anyways, as mentioned in my previous commit, I moved the prompt-prefixed functions to the interface-related for reasons mentioned in that said commit.

Added another function to this file, new comments on undocumented function prototypes, and improved the shiftPosition function. (which was troublesome to implement since 2 weeks ago)